+++ defeatbeta_api/data/ticker.py
@@ -901,7 +901,7 @@ class Ticker:
 
             merged_df = pd.merge_asof(
                 ttm_net_income_df[['report_date', symbol]].rename(columns={symbol: 'ttm_net_income'}),
-                currency_df.sort_values('report_date'),
+                currency_df,
                 left_on='report_date',
                 right_on='report_date',
                 direction='backward'
@@ -925,6 +925,84 @@ class Ticker:
             )
 
         df['industry_pe'] = (df['total_market_cap'] / df['total_ttm_net_income']).replace([np.inf, -np.inf], np.nan).round(2)
+        return df
+
+    def industry_ps_ratio(self) -> pd.DataFrame:
+        info = self.info()
+        industry = info['industry']
+        if isinstance(industry, pd.Series):
+            industry = industry.iloc[0]
+
+        if not industry or pd.isna(industry):
+            raise ValueError(f"Unknown industry for this ticker: {self.ticker}")
+
+        url = self.huggingface_client.get_url_path(stock_profile)
+        sql = load_sql("select_tickers_by_industry", url=url, industry=industry)
+        symbols = self.duckdb_client.query(sql)['symbol']
+        symbols = symbols[symbols != self.ticker]
+        symbols = pd.concat([pd.Series([self.ticker]), symbols], ignore_index=True)
+
+        market_cap_table_sql = load_sql("select_market_cap_by_industry",
+                                        stock_prices=self.huggingface_client.get_url_path(stock_prices),
+                                        stock_shares_outstanding=self.huggingface_client.get_url_path(
+                                            stock_shares_outstanding),
+                                        symbols=", ".join(f"'{s}'" for s in symbols))
+
+        total_market_cap = self.duckdb_client.query(market_cap_table_sql)
+
+        market_cap_cols = [col for col in total_market_cap.columns if col != 'report_date']
+
+        total_market_cap['total_market_cap'] = total_market_cap[market_cap_cols].sum(axis=1, skipna=True)
+        total_market_cap['industry'] = industry
+        total_market_cap = total_market_cap[['report_date', 'industry', 'total_market_cap']]
+        total_market_cap['report_date'] = pd.to_datetime(total_market_cap['report_date'])
+
+        ttm_revenue_sql = load_sql("select_ttm_revenue_by_industry",
+                                      stock_statement = self.huggingface_client.get_url_path(stock_statement),
+                                      symbols = ", ".join(f"'{s}'" for s in symbols))
+        ttm_revenue = self.duckdb_client.query(ttm_revenue_sql)
+        ttm_revenue_df = ttm_revenue.copy()
+        currency_dict = load_financial_currency()
+        ttm_revenue_df['report_date'] = pd.to_datetime(ttm_revenue_df['report_date'])
+        for symbol in ttm_revenue_df.columns:
+            if symbol == 'report_date':
+                continue
+            currency = currency_dict.get(symbol, 'USD')
+            if currency == 'USD':
+                currency_df = pd.DataFrame({
+                    'report_date': ttm_revenue_df['report_date'],
+                    'close': 1.0
+                })
+            else:
+                currency_df = self.currency(symbol=currency + '=X')
+                currency_df['report_date'] = pd.to_datetime(currency_df['report_date'])
+
+            merged_df = pd.merge_asof(
+                ttm_revenue_df[['report_date', symbol]].rename(columns={symbol: 'ttm_revenue'}),
+                currency_df,
+                left_on='report_date',
+                right_on='report_date',
+                direction='backward'
+            )
+            ttm_revenue_df[f'{symbol}_usd'] = (merged_df['ttm_revenue'] / merged_df['close']).round(2)
+
+        cols_to_keep = ['report_date'] + [c for c in ttm_revenue_df.columns if c.endswith('_usd')]
+        ttm_revenue_df = ttm_revenue_df[cols_to_keep]
+        ttm_revenue_df = ttm_revenue_df.ffill()
+
+        ttm_revenue_df_cols = [col for col in ttm_revenue_df.columns if col != 'report_date']
+        ttm_revenue_df['total_ttm_revenue'] = ttm_revenue_df[ttm_revenue_df_cols].sum(axis=1, skipna=True)
+        ttm_revenue_df = ttm_revenue_df[['report_date', 'total_ttm_revenue']]
+        ttm_revenue_df['report_date'] = pd.to_datetime(ttm_revenue_df['report_date'])
+        df = pd.merge_asof(
+                total_market_cap,
+                ttm_revenue_df,
+                left_on='report_date',
+                right_on='report_date',
+                direction='backward'
+            )
+
+        df['industry_ps_ratio'] = (df['total_market_cap'] / df['total_ttm_revenue']).replace([np.inf, -np.inf], np.nan).round(2)
         return df
 
     def _quarterly_eps_yoy_growth(self, eps_column: str, current_alias: str, prev_alias: str) -> pd.DataFrame: